backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.3",
    "anyio>=4.6.0",
    "asyncpg>=0.29.0",
    "fastapi-pagination>=0.12.31",
    "fastapi[standard]>=0.115.2",
    "faststream>=0.5.26",
    "geoalchemy2[shapely]>=0.15.2",
    "googlemaps>=4.10.0",
    "ipython>=8.28.0",
    "isort>=5.13.2",
    "nanoid>=2.0.0",
    "psycopg>=3.2.3",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.12",
    "redis>=5.1.1",
    "setuptools>=75.2.0",
    "sqlalchemy>=2.0.35",
    "sse-starlette>=2.1.3",
]
